AutoPanel Cycle Ready
Simulation happens in the cycle. Added a boolean to determine if the
simulation is on, as well as stand value and bet value for the
simulation. Those values are set in the constructor by getting values
from properties.

Did have to add a getter from gameBoard to get access to player's points
(for comparison with standValue). Also, looks like we can scrap the
progress bar because we already have it as a method.